Set HAMA to pull series 'Average Rating'
This is purely personal preference, but I have disliked AniDB's 'Weighted Rating' system since it was introduced - this commit makes HAMA pull the plain average rating instead.

As for why the variables are called "temporary" and "permanent", the way AniDB works is that, while a show airs, it can only be given a temporary rating, and after it has finished airing, it can be assigned either a temporary or permanent rating, depending on whether you have completed it or not. In addition, AniDB shows both the weighted rating, designed to redistribute ratings so that they span the full 1-10 scale, and the plain average rating, calculated by taking all ratings and dividing by the number of ratings.

But when you look at the XML, you only see two ratings, temporary and permanent; those variable names come from before the restructuring of the system, and are no longer properly representative of the value they store: While a show is airing, a permanent rating cannot be assigned, so the blank variable is used to store the weighted rating. The average of the temporary ratings is assigned to the temporary variable. 

Once a show has finished airing, the permanent variable is used to store the weighted rating, based on all permanent ratings.  The temporary variable, on the other hand, is used to store the plain average of ALL ratings of the series, permanent and temporary. This saved them from having to add two new variables to the API, two of which would have been NULL until the series finished airing. 

As mentioned, this is strictly personal preference, but I find the average rating to be closer to how I would rate a series, regardless of it not properly using the full scale: to me, ratings above 5 represent series with at least some redeeming quality, and below 5 represent truly bad shows, the lower, the worse (essentially a -5 to 5 scale), and in my experience, most people seem to vote with that mentality, with 7 generally being seen as "average".